Move the permissionTest folder inside of test folder.
