--- permissionTest/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/cloud-barista/cb-spider/permissionTest/cloudTest"
-
-func main(){
-	cloudTest.RunAllTests()
-}
