+++ test/permissionTest/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/cloud-barista/cb-spider/test/permissionTest/cloudTest"
+
+func main(){
+	cloudTest.RunAllTests()
+}
