+++ test/permissionTest/resources/keypair.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/cloud-barista/cb-spider/permissionTest/connection"
-	"github.com/cloud-barista/cb-spider/permissionTest/request"
+	"github.com/cloud-barista/cb-spider/test/permissionTest/connection"
+	"github.com/cloud-barista/cb-spider/test/permissionTest/request"
 )
 
 type KeypairResource struct{}
