+++ test/permissionTest/resources/vpc.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/cloud-barista/cb-spider/permissionTest/connection"
-	"github.com/cloud-barista/cb-spider/permissionTest/request"
+	"github.com/cloud-barista/cb-spider/test/permissionTest/connection"
+	"github.com/cloud-barista/cb-spider/test/permissionTest/request"
 )
 
 type VPCResource struct{}
